// packages/api_contracts/generated.ts
export interface components {
    schemas: {
// ...
        BusinessEvent: {
            /** Format: uuid */
            eventId: string;
            /** @constant */
            type: "lesson.completed.v1";
            /** @constant */
            schemaVersion: 1;
            /** @constant */
            producer: "learning";
            aggregateId: string;
            /** Format: date-time */
            occurredAt: string;
            requestId: string;
            payload: components["schemas"]["LessonEventPayload"];
        } | {
            /** Format: uuid */
            eventId: string;
            /** @constant */
            type: "account.deleted.v1";
            /** @constant */
            schemaVersion: 1;
            /** @constant */
            producer: "progress";
            aggregateId: string;
            /** Format: date-time */
            occurredAt: string;
            requestId: string;
            payload: components["schemas"]["DeletionEventPayload"];
        } | {
            /** Format: uuid */
            eventId: string;
            /** @constant */
            type: "account.purged.v1";
            /** @constant */
            schemaVersion: 1;
            /** @enum {string} */
            producer: "learning" | "ai";
            aggregateId: string;
            /** Format: date-time */
            occurredAt: string;
            requestId: string;
            payload: components["schemas"]["PurgeEventPayload"];
        } | {
            /** Format: uuid */
            eventId: string;
            /** @constant */
            type: "zone.changed.v1";
            /** @constant */
            schemaVersion: 1;
            /** @constant */
            producer: "progress";
            aggregateId: string;
            /** Format: date-time */
            occurredAt: string;
            requestId: string;
            payload: components["schemas"]["ZoneEventPayload"];
        };
// ...
    };
// ...
}